Create docker-compose.yml
Compose file to launch postgres, redis, and Flower along with the Airflow webserver, workers, and scheduler container. The environment variables require compose version >= 3.4. This builds the image from the Dockerfile in the same directory and passes the --build-args directly in the compose file. Of course, this can be changed to point to a pre-built image as well.